Implement platform specific disc stats
