+++ internal/stats/diskinfo/diskinfo.go
@@ -1,10 +1,9 @@
-package stats
+package diskinfo
 
 import (
 	"fmt"
 	"github.com/NodeFactoryIo/hactar-daemon/internal/hactar"
 	log "github.com/sirupsen/logrus"
-	"golang.org/x/sys/unix"
 	"net/http"
 )
 
@@ -31,17 +30,4 @@ func SendDiskInfoStats(client *hactar.Client, actorAddress string, nodeUrl strin
 
 	log.Error("Unable to send disk information statistics. ", err)
 	return false
-}
-
-// calculates disk usage of path/disk
-func DiskUsage(path string) (disk DiskStatus) {
-	fs := unix.Statfs_t{}
-	err := unix.Statfs(path, &fs)
-	if err != nil {
-		return
-	}
-	disk.All = fs.Blocks * uint64(fs.Bsize)
-	disk.Free = fs.Bfree * uint64(fs.Bsize)
-	disk.Used = disk.All - disk.Free
-	return
 }
